rename "inspector.tsx" -> "controls.tsx"
